movies files turned into more generic
